gn: Add gn build infra
Setting up the infra for building the clspv targets using GN[1]. The
build is heavily dependent on the LLVM GN build infra and leverages it
as such.

The build framework is based on the LLVM build files. The files are
referenceed to the original LLVM wherever possible, and few of the files
are copied with modifications on paths.

This is squashed version of branch
`dev/build-from-angle-with-in-tree-gn-work`

Signed-off-by: Gowtham Tammana <g.tammana@samsung.com>
Change-Id: I1ea822eab9ab164cb076fa96150c580ba53d485f