gn: Add gn build infra
Setting up the infra for building the clspv targets using GN[1]. The
build is heavily dependent on the LLVM GN build infra and leverages it
as such.

The GN files are setup to be in *secondary_source* location similar to
that of LLVM.

The build currently requires additional step of symlinking the LLVM gn
files to ensure source tree of llvm is matched to this project.

Signed-off-by: Gowtham Tammana <g.tammana@samsung.com>
Change-Id: I1ea822eab9ab164cb076fa96150c580ba53d485f